Fixes for not provided log variable
+ implemented methods:
- validateJson
- store
- update

+++ includes/zenci-log-delete.js
@@ -66,10 +66,12 @@ LogDelete.prototype.process = function( callback ) {
               }
             } );
           } else {
-            self.fileDir = process.env.FILE_DIR + "/" + result.value.owner +
-              "/" + result.value.repository;
-            if ( fs.existsSync( self.fileDir + "/" + self.requestDetails.url ) ) {
-              fs.unlink( self.fileDir + "/" + self.requestDetails.url );
+            if(process.env.FILE_DIR) {
+              self.fileDir = process.env.FILE_DIR + "/" + result.value.owner +
+                "/" + result.value.repository;
+              if ( fs.existsSync( self.fileDir + "/" + self.requestDetails.url ) ) {
+                fs.unlink( self.fileDir + "/" + self.requestDetails.url );
+              }
             }
             callback( null, {
               code: 200,
